+++ lib/crudrepo.dart
@@ -8,10 +8,10 @@
 
 library;
 
-export 'src/dio_service.dart';
-export 'src/sqlitelib/feds_local.dart';
-export 'src/sqlitelib/feds_local_sqflite.dart';
-export 'src/sqlitelib/feds_local_sqflite_ffi.dart';
+export 'src/services/dio_service.dart';
+export 'src/services/sqlitelib/feds_local.dart';
+export 'src/services/sqlitelib/feds_local_sqflite.dart';
+export 'src/services/sqlitelib/feds_local_sqflite_ffi.dart';
 export 'src/repositories/crud_repository.dart';
 export 'src/repositories/crud_repository_local.dart';
 export 'src/repositories/crud_repository_remote.dart';