+++ lib/src/repositories/crud_repository_cached.dart
@@ -1,7 +1,7 @@
 import 'package:result_dart/result_dart.dart';
 
-import '../dio_service.dart';
-import '../sqlitelib/feds_local.dart';
+import '../services/dio_service.dart';
+import '../services/sqlitelib/feds_local.dart';
 import 'crud_repository.dart';
 
 /// A repository for performing CRUD operations on a remote API with local caching.